Remove useless casts from the emulator
Suppose an expression E has type T. Then, if E is cast to T, the cast
is useless in the sense that it does not do anything at best. At
worst, however the cast may hide an error.

This patch removes all useless casts from the ERTS emulator by means
of automatic patch generation. It uses the tool Coccinelle, a semantic
patch suite. In this case, we applied the patch:

@@
type T;
T E;
@@

-     (T)
      E

+++ erts/emulator/beam/erl_cpu_topology.c
@@ -128,7 +128,7 @@ static erts_cpu_groups_map_t *reader_groups_map;
 
 #define ERTS_TOPOLOGY_CG ERTS_TOPOLOGY_MAX_DEPTH
 
-#define ERTS_MAX_CPU_TOPOLOGY_ID ((int) 0xffff)
+#define ERTS_MAX_CPU_TOPOLOGY_ID (0xffff)
 
 #ifdef ERTS_SMP
 static void cpu_bind_order_sort(erts_cpu_topology_t *cpudata,
